// src/types.ts
export type LanguageCode = 
  | 'te' 
  | 'hi' 
  | 'en' 
  | 'ta' 
  | 'kn' 
  | 'ml' 
  | 'mr' 
  | 'bn' 
  | 'gu' 
  | 'pa' 
  | 'or' 
  | 'as' 
  | 'ur';
// ...
export interface SystemStats {
  total_complaints: number;
  complaints_resolved: number;
  active_complaints: number;
  pending_verification: number;
  average_resolution_hours: number;
  sla_compliance_rate: number;
  departments_connected: number;
  incident_clusters_active: number;
  citizen_satisfaction_rate: number;
}